+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Support multiple varfiles when create/updating variables. [#28]
+
 ### Fixed
 
 * Incorrect enconding of HCL lists. [#27]
@@ -51,3 +55,4 @@ Initial version with support for managing:
 [#20]: https://github.com/rgreinho/tfe-cli/pull/20
 [#23]: https://github.com/rgreinho/tfe-cli/pull/23
 [#27]: https://github.com/rgreinho/tfe-cli/pull/27
+[#28]: https://github.com/rgreinho/tfe-cli/pull/28